+++ characterFiles/Attribute.java
@@ -1,6 +1,8 @@
 package characterFiles;
 
 /**
+ * Version: 0.1.Alpha 2020/08/27 Steven Anderson
+ * 
  * The Attribute is the basis of all character stats. Most character features
  * are based from the Attribute Class.
  */
@@ -8,23 +10,27 @@ package characterFiles;
 public class Attribute {
     private String attributeName;
     private int attributeScore;
-    private int attributeModifier;
+    private int attributeModifier = 0;
 
     // Create an attribute taking a String and int, the modifier is calculated.
     public Attribute(String attributeName, int attributeScore) {
         this.attributeName = attributeName;
         this.attributeScore = attributeScore;
-        attributeModifier = calculateModifier(attributeScore);
+        updateModifier();
     }
 
-    // Calculates modifier based on score, if the modifier is >=10 the result is
-    // positive, otherwise negative
-    private int calculateModifier(int score) {
-        if (score >= 10) {
-            return (int) Math.floor((score - 10) / 2);
-        } else {
-            return (int) Math.floor((score - 10.1) / 2);
-        }
+    public void setScore(int newScore) {
+        attributeScore = newScore;
+        updateModifier();
+    }
+
+    /**
+     * Update the modifier. Call this each time the score changes.
+     */
+    private void updateModifier() {
+        // Calculates modifier based on score, if the modifier is >=10 the result is
+        // positive, otherwise negative
+        attributeModifier = Math.floorDiv((attributeScore - 10),2);
     }
 
     public String getAttributeName() {
@@ -40,6 +46,6 @@ public class Attribute {
     }
 
     public String toString() {
-        return String.format("%s [%s] %s%n", attributeName, attributeScore, attributeModifier);
+        return String.format("%s [%s] %s%n", getAttributeName(), getAttributeScore(), getAttributeModifier());
     }
 }